requirements.txt
numpy==1.23.5
Pillow==9.5.0
requests==2.25.1
pymongo==3.11.3
Flask==2.3.3
pytest==6.2.2
python-dotenv
psycopg2-binary